refactor around dataset superclass
Moved all complexity to a small superclass and set
of utility functions so that subclasses have only
the essentials that make them different.

Still maintain the dataset.load_data() static
function as the primary interface, which simply
delegates to an instantiated class.

Added dataset versioning with version string
being part of the filename, which involved some
light fuel monkey-patching of classes.